Extract the useTasksSelected hook
